Fixed bug with Unification.SINGLETON
Before, it was a SINGLE object returned from LazySequence.of(). That had
the effect that the first time tryAdvance() was called on the SINGLETON,
it worked as expected. If then, anywhere else in the code, SINGLETON
was used again, it acted the same as Unification.NONE; that is
obviously incorrect.
The solution was to define a getter so a new LazySequence is created
every time SINGLETON is accessed.